Change to MIT license.
I don't think there's any difference, really. I just noticed that
github now provides an option to create a license when creating
a new repo, and in the select box, the MIT license is shown in bold,
whereas the BSD license is not. So that suggests that the MIT license
must be superior in some way. And that's good enough for me.

(I think software licenses have been overcomplicated.)