Keep moving stuff around.
I want to get all the classes together, near the top of the module.
Mainly, I keep forgetting which terms are classes and which ones
are functions. I want to be able to look at all the classes together.
So that's why I'm shuffling all this stuff around (and probably
totally messing up the history - oh well).